Optimize timestamp setting in AuditedDbContext
The code changes optimize the setting of `UpdatedAtUtc` and `CreatedAtUtc` properties in the `AuditedDbContext` class. Instead of calling `DateTime.UtcNow.DropMicroseconds()` multiple times, the current UTC time is stored in a variable `now` and reused. This reduces the number of calls to `DateTime.UtcNow`, improving performance and ensuring consistency in the timestamps.