Add AddRange method to ClaimBuilder for bulk claim addition
A new method `AddRange` has been added to the `ClaimBuilder` class
within the `EasyExtensions.Authorization.Builders` namespace. This
method allows for adding a collection of claims to the builder. It
includes null checks for the collection itself as well as for each
individual claim and their `Type` and `Value` properties. If any of
these are null or empty, an `ArgumentNullException` is thrown. The
method returns the current instance of `ClaimBuilder`.