Update XML docs and default value for GetAnyFont method
Enhanced XML documentation for the GetAnyFont method in the
FontHelpers class. Updated the <param> tag for the size
parameter to indicate a default font size of 12. The <returns>
tag now specifies that the method returns a Font. Modified the
method signature to set a default value of 12 for the size
parameter.