Add StartNow property to JobTriggerAttribute
Added a new property `StartNow` to the `JobTriggerAttribute` class to indicate whether the job should start immediately. Updated the constructor to include a `startNow` parameter with a default value of `true`. Modified the job scheduling logic in `ServiceCollectionExtensions` to check the `StartNow` property and call `opts.StartNow()` if it is set to `true`.