Make jobAdded parameter optional in Quartz setup methods
The method `AddQuartzJobs` in `ServiceCollectionExtensions.cs` now
provides a default value of `null` for the `jobAdded` parameter,
allowing it to be called without explicitly passing an `Action<Type>`
argument. Similarly, the `SetupQuartz` method has been updated to
provide a default value of `null` for its `jobAdded` parameter. These
changes enhance the flexibility of these methods, allowing them to be
invoked with or without the `jobAdded` argument.