Add RepeatForever property to JobTriggerAttribute
Added a new property `RepeatForever` to the `JobTriggerAttribute` class to allow jobs to be optionally repeated forever. Updated the constructor of `JobTriggerAttribute` to include the new `repeatForever` parameter, with a default value of `true`. Modified the `ServiceCollectionExtensions` class to conditionally call `RepeatForever()` on the job schedule based on the value of the `RepeatForever` property in `JobTriggerAttribute`.