Remove folder validation in GetFilesAsync/GetDirectoriesAsync
The validation logic that checked if the specified folder is
actually a file has been removed from both the GetFilesAsync
and GetDirectoriesAsync methods. Previously, an
InvalidOperationException would be thrown if the folder path
pointed to a file. With this change, the methods will no
longer perform this validation and will simply return the
resources filtered by their type (files or directories).