Make setup parameter optional in UseSentryWithUserCapturing
The method signature of UseSentryWithUserCapturing in the
WebHostBuilderExtensions.cs file has been modified. Specifically,
the setup parameter now has a default value of null. This change
allows the method to be called without explicitly passing a value
for the setup parameter, simplifying the method call in cases
where no setup action is needed.