Add contextLifetime parameter to AddPostgresDbContext
The method `AddPostgresDbContext` in `ServiceCollectionExtensions.cs`
now includes an additional parameter, `contextLifetime`, which specifies
the `ServiceLifetime` of the `DbContext`. The default value for this
parameter is set to `ServiceLifetime.Scoped`.

The method signature has been updated to accept this new parameter, and
the call to `AddDbContext<TContext>` has been modified to pass the
`contextLifetime` parameter.

Additionally, the XML documentation for the method has been updated to
include a description of the new `contextLifetime` parameter.