Refactor BaseRepository and IRepository interfaces
Modified SaveChangesAsync in BaseRepository to accept a CancellationToken and changed its visibility to private. Updated several methods in BaseRepository to be virtual for better extensibility. Added new async methods CountAsync, FirstAsync, and FirstOrDefaultAsync to BaseRepository and IRepository with support for CancellationToken. Removed SaveChangesAsync from IRepository and marked Query as obsolete.